Move debugLog to helpers and update imports
