+++ .changeset/wild-coats-prove.md
@@ -1,0 +1,5 @@
+---
+"@iqai/adk": patch
+---
+
+Adds debug logs
