+++ packages/adk/src/agents/lang-graph-agent.ts
@@ -1,4 +1,4 @@
-import { debugLog } from "@adk/lib/debug";
+import { debugLog } from "@adk/helpers/debug";
 import type {
 	Message,
 	MessageContent,
