+++ packages/adk/src/models/llm-registry.ts
@@ -1,4 +1,4 @@
-import { debugLog } from "@adk/lib/debug";
+import { debugLog } from "@adk/helpers/debug";
 import type { BaseLLM } from "./base-llm";
 
 /**
